Make LinearRegression an adapted version of LeastSquaresRegression.
